feat: improve pagination block
