+++ src/shared/components/ncTable/sections/CustomTable.vue
@@ -8,7 +8,7 @@
 			<thead class="tables-list__thead">
 				<TableHeader :columns="columns"
 					:selected-rows="selectedRows"
-					:rows="getSearchedAndFilteredAndSortedRows"
+					:rows="rows"
 					:view-setting.sync="localViewSetting"
 					:config="config"
 					:pinned-column-id="pinnedColumnId"
@@ -18,6 +18,7 @@
 					@edit-column="col => $emit('edit-column', col)"
 					@delete-column="col => $emit('delete-column', col)"
 					@download-csv="data => $emit('download-csv', data)"
+
 					@select-all-rows="selectAllRows"
 					@pin-column="setPinnedColumn">
 					<template #actions>
@@ -40,70 +41,21 @@
 					:config="config"
 					:element-id="elementId"
 					:is-view="isView"
+
 					:pinned-column-id="pinnedColumnId"
 					:column-widths="columnWidths"
 					@update-row-selection="updateRowSelection"
 					@edit-row="rowId => $emit('edit-row', rowId)" />
 			</transition-group>
 		</table>
-		<div v-if="totalPages > 1" class="pagination-footer" :class="{'large-width': !appNavCollapsed || isMobile}">
-			<div class="pagination-items">
-				<NcButton type="tertiary" :disabled="totalPages === 1 || pageNumber <= 1" :aria-label="t('tables', 'Go to first page')" @click="pageNumber = 1">
-					<template #icon>
-						<PageFirstIcon :size="20" />
-					</template>
-				</NcButton>
-				<NcButton type="tertiary" :disabled="totalPages === 1 || pageNumber <= 1" :aria-label="t('tables', 'Go to previous page')" @click="pageNumber--">
-					<template #icon>
-						<ChevronLeftIcon :size="20" />
-					</template>
-				</NcButton>
-				<div class="page-number">
-					<NcSelect
-						v-model="pageNumber"
-						:options="allPageNumbersArray"
-						:clearable="false"
-						:aria-label-combobox="t('tables', 'Page number')">
-						<template #selected-option-container="{ option }">
-							<span class="selected-page">
-								{{ option.label }} of {{ totalPages }}
-							</span>
-						</template>
-					</NcSelect>
-				</div>
-				<NcButton type="tertiary" :disabled="totalPages === 1 || pageNumber >= totalPages" :aria-label="t('tables', 'Go to next page')" @click="pageNumber++">
-					<template #icon>
-						<ChevronRightIcon :size="20" />
-					</template>
-				</NcButton>
-				<NcButton type="tertiary" :disabled="totalPages === 1 || pageNumber >= totalPages" :aria-label="t('tables', 'Go to last page')" @click="pageNumber = totalPages">
-					<template #icon>
-						<PageLastIcon :size="20" />
-					</template>
-				</NcButton>
-			</div>
-		</div>
 	</div>
 </template>
 
 <script>
 import TableHeader from '../partials/TableHeader.vue'
-import PageLastIcon from 'vue-material-design-icons/PageLast.vue'
-import PageFirstIcon from 'vue-material-design-icons/PageFirst.vue'
-import ChevronRightIcon from 'vue-material-design-icons/ChevronRight.vue'
-import ChevronLeftIcon from 'vue-material-design-icons/ChevronLeft.vue'
 import TableRow from '../partials/TableRow.vue'
 import { subscribe, unsubscribe } from '@nextcloud/event-bus'
-import { MagicFields } from '../mixins/magicFields.js'
-import { NcButton, useIsMobile, NcSelect } from '@nextcloud/vue'
-import { mapState } from 'pinia'
-import {
-	TYPE_META_ID, TYPE_META_CREATED_BY, TYPE_META_CREATED_AT, TYPE_META_UPDATED_BY, TYPE_META_UPDATED_AT,
-} from '../../../constants.ts'
-import { MetaColumns } from '../mixins/metaColumns.js'
 import { translate as t } from '@nextcloud/l10n'
-import { useTablesStore } from '../../../../store/store.js'
-import { getFiltersForColumn } from '../mixins/filter.js'
 
 export default {
 	name: 'CustomTable',
@@ -111,12 +63,6 @@ export default {
 	components: {
 		TableRow,
 		TableHeader,
-		NcButton,
-		PageLastIcon,
-		PageFirstIcon,
-		ChevronLeftIcon,
-		ChevronRightIcon,
-		NcSelect,
 	},
 
 	props: {
@@ -146,12 +92,6 @@ export default {
 		},
 	},
 
-	setup() {
-		return {
-			isMobile: useIsMobile(),
-		}
-	},
-
 	data() {
 		return {
 			selectedRows: [],
@@ -166,164 +106,8 @@ export default {
 	},
 
 	computed: {
-		...mapState(useTablesStore, ['appNavCollapsed']),
-		allPageNumbersArray() {
-			return Array.from(
-				{ length: this.totalPages },
-				(value, index) => 1 + index,
-			)
-		},
 		currentPageRows() {
-			return this.getSearchedAndFilteredAndSortedRows.slice((this.pageNumber - 1) * this.rowsPerPage, ((this.pageNumber - 1) * this.rowsPerPage) + this.rowsPerPage)
-		},
-		totalPages() {
-			return Math.ceil(this.getSearchedAndFilteredAndSortedRows.length / this.rowsPerPage)
-		},
-		sorting() {
-			return this.viewSetting?.sorting
-		},
-		getSearchedAndFilteredRows() {
-			const debug = false
-			// if we don't have to search and/or filter
-			if (!this.viewSetting?.filter?.length > 0 && !this.viewSetting?.searchString) {
-				// cleanup markers
-				if (this.rows && this.columns) {
-					this.rows.forEach(row => {
-						if (row && row.data) {
-							this.columns.forEach(column => {
-								const cell = row.data.find(item => item && item.columnId === column.id)
-								if (cell === undefined) {
-									return
-								}
-								delete cell.searchStringFound
-								delete cell.filterFound
-							})
-						}
-					})
-				}
-				return this.rows || []
-			}
-
-			const data = [] // array of rows
-			const searchString = this.viewSetting?.searchString
-			// each row
-			if (!this.rows || !this.columns) {
-				return []
-			}
-
-			for (const row of this.rows) {
-				if (!row || !row.data) {
-					continue
-				}
-
-				if (debug) {
-					console.debug('new row ===============================================', row)
-				}
-				let filterStatusRow = null
-				let searchStatusRow = false
-
-				// each column in a row => cell
-				this.columns.forEach(column => {
-					if (debug) {
-						console.debug('new column -------------------', column)
-					}
-					let filterStatus = null
-					let searchStatus = true
-					const filters = getFiltersForColumn(column, this.viewSetting)
-					let cell
-					if (column.id < 0) {
-						cell = { columnId: column.id }
-						switch (column.id) {
-						case TYPE_META_ID:
-							cell.value = row.id
-							break
-						case TYPE_META_CREATED_BY:
-							cell.value = row.createdBy
-							break
-						case TYPE_META_UPDATED_BY:
-							cell.value = row.editedBy
-							break
-						case TYPE_META_CREATED_AT:
-							cell.value = row.createdAt
-							break
-						case TYPE_META_UPDATED_AT:
-							cell.value = row.editedAt
-							break
-						}
-					} else {
-						cell = row.data.find(item => item && item.columnId === column.id)
-					}
-
-					// if we don't have a value for this cell
-					if (cell === undefined) {
-						if (searchString) {
-							searchStatus = false
-						}
-						cell = { columnId: column.id, value: null }
-					}
-					// cleanup possible old markers
-					delete cell.searchStringFound
-					delete cell.filterFound
-
-					// apply filters (if any)
-					filters.forEach(fil => {
-						this.addMagicFieldsValues(fil)
-						if (filterStatus === null || filterStatus === true) {
-							filterStatus = column.isFilterFound(cell, fil)
-						}
-					})
-					// if we should search
-					if (searchString) {
-						console.debug('look for searchString', searchString)
-						searchStatus = column.isSearchStringFound(cell, searchString.toLowerCase())
-					}
-
-					if (debug) {
-						console.debug('filterStatus for cell', { cell: cell?.value, filterStatusCell: filterStatus, filterStatusRowBefore: filterStatusRow })
-					}
-
-					// if filterStatus is null, this result should be ignored
-					if (filterStatus !== null && (filterStatusRow || filterStatusRow === null)) {
-						filterStatusRow = filterStatus
-					}
-
-					if (debug) {
-						console.debug('new filterStatusRow', filterStatusRow)
-					}
-
-					// filterStatusRow = filterStatus
-					searchStatusRow = searchStatusRow || searchStatus
-				})
-
-				if (debug) {
-					console.debug('if push row', { filterStatusRow, searchStatusRow, result: (filterStatusRow || filterStatusRow === null) && searchStatusRow })
-				}
-				if ((filterStatusRow || filterStatusRow === null) && searchStatusRow) {
-					data.push({ ...row })
-				}
-			}
-			return data
-		},
-		getSearchedAndFilteredAndSortedRows() {
-			const allColumns = this.columns.concat(MetaColumns)
-			const sort = (sortCols) => {
-				const sortColumn = allColumns.find(item => item.id === sortCols?.[0].columnId)
-				const nextSorts = []
-				for (let i = 1; i < sortCols.length; i++) {
-					const sortColumn = allColumns.find(item => item.id === sortCols[i].columnId)
-					nextSorts.push(sortColumn?.sort?.(sortCols[i].mode))
-				}
-				return [...this.getSearchedAndFilteredRows].sort(sortColumn?.sort?.(sortCols[0].mode, nextSorts))
-			}
-
-			// if we have to sort
-			if (this.viewSetting?.presetSorting) {
-				return sort(this.viewSetting.presetSorting)
-			}
-			if (this.viewSetting?.sorting) {
-				return sort(this.viewSetting.sorting)
-			}
-			return this.getSearchedAndFilteredRows
+			return this.rows.slice((this.pageNumber - 1) * this.rowsPerPage, ((this.pageNumber - 1) * this.rowsPerPage) + this.rowsPerPage)
 		},
 	},
 
@@ -343,19 +127,15 @@ export default {
 		},
 	},
 
-	updated() {
-		if (this.pageNumber > this.totalPages || this.totalPages === 1) {
-			this.pageNumber = this.totalPages
-		}
-	},
-
 	mounted() {
 		subscribe('tables:selected-rows:deselect', ({ elementId, isView }) => this.deselectAllRows(elementId, isView))
 		subscribe('tables:row:animate', this.enableRowAnimation)
+		subscribe('tables:pagination-changed', this.handlePaginationChanged)
 	},
 	beforeDestroy() {
 		unsubscribe('tables:selected-rows:deselect', ({ elementId, isView }) => this.deselectAllRows(elementId, isView))
 		unsubscribe('tables:row:animate', this.enableRowAnimation)
+		unsubscribe('tables:pagination-changed', this.handlePaginationChanged)
 	},
 
 	methods: {
@@ -374,13 +154,11 @@ export default {
 				this.columnWidths = widths
 			}
 		},
-		addMagicFieldsValues(filter) {
-			Object.values(MagicFields).forEach(field => {
-				const newFilterValue = filter.value.replace('@' + field.id, field.replace)
-				if (filter.value !== newFilterValue) {
-					filter.magicValuesEnriched = newFilterValue
-				}
-			})
+		handlePaginationChanged({ pageNumber, rowsPerPage }) {
+			this.pageNumber = pageNumber
+			if (rowsPerPage) {
+				this.rowsPerPage = rowsPerPage
+			}
 		},
 		deselectAllRows(elementId, isView) {
 			if (parseInt(elementId) === parseInt(this.elementId) && isView === this.isView) {
@@ -390,7 +168,7 @@ export default {
 		selectAllRows(value) {
 			this.selectedRows = []
 			if (value) {
-				this.getSearchedAndFilteredRows.forEach(item => { this.selectedRows.push(item.id) })
+				this.rows.forEach(item => { this.selectedRows.push(item.id) })
 			}
 			this.$emit('update-selected-rows', this.selectedRows)
 		},
@@ -423,12 +201,6 @@ export default {
 }
 </script>
 
-<style>
-.vs__dropdown-menu {
-	min-width: 95px !important;
-}
-</style>
-
 <style lang="scss" scoped>
 :deep(.text-editor__wrapper .paragraph-content:last-child) {
 	margin-bottom: 0!important;
@@ -436,47 +208,6 @@ export default {
 
 :deep(.text-editor__wrapper .ProseMirror > *:first-child) {
 	margin-top: 0!important;
-}
-
-.selected-page{
-	padding-inline-start: 5px;
-
-	display:inline-flex;
-	align-items: center;
-}
-
-.page-number{
-	padding-inline: 5px;
-}
-
-.large-width{
-	width: 100vw !important;
-	inset-inline-start: 0 !important;
-}
-
-.pagination-items{
-	background-color: var(--color-main-background);
-	border-radius: var(--border-radius-large);
-	pointer-events: all;
-
-	display: flex;
-	align-items: center;
-}
-
-.pagination-footer{
-	box-shadow: var(--box-shadow);
-	filter: drop-shadow(0 1px 6px var(--color-box-shadow));
-	padding-bottom: 20px;
-	width: calc(100vw - 316px);
-	pointer-events: none;
-
-	display: flex;
-	justify-content: center;
-	align-items: center;
-
-	:deep(.v-select) {
-		min-width: 95px !important;
-	}
 }
 
 :deep(table) {
